requirements.txt
asgiref==3.4.1
comtypes==1.1.10
Django==3.2.9
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
install==1.3.4
Pillow==8.4.0
pydub==0.25.1
pypiwin32==223
pyttsx3==2.90
pytz==2021.3
pywin32==302
SpeechRecognition==3.8.1
sqlparse==0.4.2
